+++ src/test/java/de/wwu/scdh/teilsp/config/ExtensionConfigurationReaderTest.java
@@ -21,11 +21,13 @@ public class ExtensionConfigurationReaderTest {
     InputStream input;
     List<ExtensionConfiguration> configs;
     ExtensionConfiguration config;
+    EditorVariablesExpander expander;
     
     @BeforeEach
     void setup() throws FileNotFoundException, IOException {
 	String configFile = Paths.get("src", "test", "resources", "config.xml").toFile().getAbsolutePath();
 	input = new FileInputStream(configFile);
+	expander = new NoExpander();
     }
 
     @AfterEach
@@ -35,7 +37,7 @@ public class ExtensionConfigurationReaderTest {
     
     @Test
     void testConfigReader() throws ConfigurationException {
-	configs = ExtensionConfigurationReader.getExtensionsConfiguration(input);
+	configs = ExtensionConfigurationReader.getExtensionsConfiguration(input, expander);
 	assertEquals(1, configs.size());
 	config = configs.get(0);
 	assertEquals("de.wwu.scdh.teilsp.extensions.LabelledEntriesFromXMLByPrefixDef", config.getClassName());
@@ -51,5 +53,11 @@ public class ExtensionConfigurationReaderTest {
 	config = configs.get(0);
 	assertEquals(4, config.getSpecification().get(1).getConditions().size());
     }
-    
+
+    @Test
+    void testPropertyReader() throws ConfigurationException {
+	String property = ExtensionConfigurationReader.getProperty("test.property", input);
+	assertEquals("TestPropertyValue", property);
+    }
+
 }